Two-Stage Path Analysis

Fit a two-stage path analysis (2S-PA) model.

Usage

tspa(model, data, reliability = NULL, se = NULL, ...)

Arguments

model

A string variable describing the structural path model, in lavaan syntax.

data

A data frame containing factor scores.

reliability

A numeric vector representing the reliability indexes of each latent factor. Currently tspa() does not support the reliability argument. Please use se.

se

A numeric vector representing the standard errors of each latent factor for single-group 2S-PA. A list or data frame storing the standard errors of each group in each latent factor for multigroup 2S-PA.

...

Additional arguments passed to sem. See lavOptions for a complete list.

Value

An object of class lavaan, with an attribute tspaModel

that contains the model syntax.

Examples

library(lavaan)
### single-group example

# get factor scores
fs_dat_ind60 <- get_fs(data = PoliticalDemocracy,
                       model = "ind60 =~ x1 + x2 + x3")
fs_dat_dem60 <- get_fs(data = PoliticalDemocracy,
                       model = "dem60 =~ y1 + y2 + y3 + y4")
fs_dat <- cbind(fs_dat_ind60, fs_dat_dem60)

# tspa model
tspa(model = "dem60 ~ ind60", data = fs_dat,
     se = c(ind60 = 0.1213615, dem60 = 0.6756472))
#> lavaan 0.6.13 ended normally after 17 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         3
#> 
#>   Number of observations                            75
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0

### multigroup example

# get factor scores
fs_dat_visual <- get_fs(data = HolzingerSwineford1939,
                        model = "visual =~ x1 + x2 + x3",
                        group = "school")
fs_dat_speed <- get_fs(data = HolzingerSwineford1939,
                       model = "speed =~ x7 + x8 + x9",
                       group = "school")
fs_hs <- cbind(fs_dat_visual, fs_dat_speed)

if (FALSE) {
# tspa model
tspa(model = "visual ~ speed",
     data = fs_hs,
     se = data.frame(visual = c(0.3391326, 0.311828),
                     speed = c(0.2786875, 0.2740507)),
     group = "school",
     group.equal = "regressions")

# manually adding equality constraints on the regression coefficients
tspa(model = "visual ~ c(b1, b1) * speed",
     data = fs_hs,
     se = list(visual = c(0.3391326, 0.311828),
               speed = c(0.2786875, 0.2740507)),
     group = "school")
}